requirements.txt
langchain
langchain-community
langchain-text-splitters
psycopg2
psycopg2-binary
python-dotenv
tiktoken
openai
pgvector
phidata

langchain-postgres
langchain-openai
